Reorganize as countertop package
This is a major reorganization of the code base, shifting away from the
previous plan of having this repository serve as a standalone package
that can be cloned, configured, and run. We have moved to an API
approach for running a copy of the TV kitchen, which will give
developers much more direct control over which components they want to
use.

This repository is now the home of the `countertop` package, which is
the central engine of the TV Kitchen, responsible for setting up work
stations, piping data between appliances, and emitting events.

This also changes the `CountertopCoordinator` to simply be `Countertop`.
This is because in the past we named it in terms of a metaphor which has
slowly phased away.

Issue #85
