+++ CHANGELOG.md
@@ -1,0 +1,11 @@
+# Changelog for @tvkitchen/countertop
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+- Initial implementation of the `countertop` package.
