+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "tv-kitchen",
-  "version": "1.0.0",
-  "description": "TV Kitchen is a playground that turns TV into data (transcripts, facial recognition, etc.) so that newsrooms can help their staff and readers better understand what is happening across various local and cable TV news outlets.",
+  "name": "@tvkitchen/countertop",
+  "version": "0.1.0",
+  "description": "The entry point for developers who want to set up a TV Kitchen.",
   "main": "index.js",
   "scripts": {
     "kafka": "yarn kafka:start",
@@ -25,7 +25,7 @@
     "journalism",
     "news"
   ],
-  "author": "Bad Idea Factory <author@biffud.com>",
+  "author": "Bad Idea Factory <biffuddotcom@biffud.com>",
   "license": "Apache-2.0",
   "bugs": {
     "url": "https://github.com/tvkitchen/tv-kitchen/issues"
