+++ src/classes/index.js
@@ -1,4 +1,4 @@
-export { default as CountertopCoordinator } from './CountertopCoordinator'
+export { default as Countertop } from './Countertop'
 export { default as CountertopStation } from './CountertopStation'
 export { default as CountertopStream } from './CountertopStream'
 export { default as CountertopTopology } from './CountertopTopology'
