--- src/constants/.gitkeep
@@ -1,1 +1,0 @@
-Delete this file once there is another item in the directory.
