+++ src/index.js
@@ -1,4 +1,1 @@
-import 'module-alias/register'
-import logger from '%src/tools/logger'
-
-logger.info('Started.')
+export * from './classes'
