+++ src/tools/utils/__test__/countertop.unit.test.js
@@ -2,7 +2,7 @@ import { by } from '%src/tools/utils'
 import {
 	CountertopStation,
 	CountertopStream,
-} from '%src/components/countertop'
+} from '%src/classes'
 import {
 	generateMockAppliance,
 	normalizeTributaryMaps,
